Kohteita voi poistaa ja päivittää, kohteen voi poistaa vasta kun se on epäaktiivinen (estetään samanaikaisuuden halllinnan ongelmat), erillinen näkymä adminille (oikeuksia ei vielä toteutettu)

+++ application/betting_offers/forms.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import BooleanField, DecimalField, IntegerField, validators
+from wtforms import BooleanField, DecimalField, validators
 
 def validate_risk(form, field):
     try:
@@ -13,7 +13,7 @@ class Betting_offerForm(FlaskForm):
     odds_1 = DecimalField("Odds for home win ", places=2, validators=[validators.NumberRange(min=1.00, max=90), validate_risk])
     odds_x = DecimalField("Odds for draw", places=2, validators=[validators.NumberRange(min=1.00, max=90), validate_risk])
     odds_2 = DecimalField("Odds for away win", places=2, validators=[validators.NumberRange(min=1.00, max=90), validate_risk])
-    max_stake = IntegerField("Maximum stake (0-100)", [validators.NumberRange(min=0,max=100)])
+    max_stake = DecimalField("Maximum stake (0-100)", [validators.NumberRange(min=0.0,max=100.0)])
     active = BooleanField("The betting offer is active")
     closed = BooleanField("The betting offer is closed")
